requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
pyaml==23.9.7
    # via python-helpers (pyproject.toml)
pyyaml==6.0.1
    # via pyaml
